services/kaizen-bridge/projector/pyproject.toml
[project]
name = "kaizen-bridge-projector"
version = "0.1.0"
dependencies = ["fastapi", "uvicorn", "pydantic", "numpy", "torch", "pyyaml"]

[tool.uvicorn]
reload = true
